[修改]注册登陆功能完善

+++ bs-user/bs-user-interface/src/main/java/cn/edu/njucm/wp/bs/user/pojo/User.java
@@ -33,7 +33,7 @@ public class User {
 
     private Integer ip;
 
-    private byte flag;
+    private Integer flag;
 
     private LocalDateTime createdAt;
